Fix the invalid permission check in `tl_content` (see #8698)
Description
-----------

fixes https://github.com/contao/contao/issues/8674

The issue is that we do not check for the parent table of the current record, only for the UNION query. If the first record has a different ptable, the recursive relation check will continue to find content elements that are related to the pid of the first record. The `tl_content` database in the dump looks like this:

<table>
<tr>
<th>id</th>
<th>pid</th>
<th>ptable</th>
<th>type</th>
</tr>
<tr>
<td>1</td>
<td>1</td>
<td>tl_article</td>
<td>element_group</td>
</tr>
<tr>
<td>2</td>
<td>1</td>
<td>tl_content</td>
<td>text</td>
</tr>
<tr>
<td>3</td>
<td>2</td>
<td> tl_article </td>
<td>element_group</td>
</tr>
<tr>
<td>4</td>
<td>3</td>
<td>tl_content</td>
<td>text</td>
</tr>
</table>

The recursive query will load a result for `id=3` and recursive to `pid` of that record (PID 2), which will incorrectly return result ID 1, which will then incorrectly check permission on the parent of ID 1 (the article with ID 1).
By checking that the original record also has `ptable=$table`, we will not load any record from the database and only check for the parent of our current record.

Commits
-------

881b051d Fix invalid permission check in tl_content
81e79681 Fix the issue in a more efficient way