Fix the `getAllEvents` hook (see #8603)
Description
-----------

If there is a `getAllEvents` hook in your system, the following error will occur:

```
TypeError:
App\EventListener\GetAllEventsListener::__invoke(): Argument #5 ($module) must be of type Contao\Module, Contao\CalendarBundle\Generator\CalendarEventsGenerator given, called in vendor\contao\contao\calendar-bundle\src\Generator\CalendarEventsGenerator.php on line 108

  at src\EventListener\GetAllEventsListener.php:12
  at App\EventListener\GetAllEventsListener->__invoke()
     (vendor\contao\contao\calendar-bundle\src\Generator\CalendarEventsGenerator.php:108)
  at Contao\CalendarBundle\Generator\CalendarEventsGenerator->getAllEvents()
     (vendor\contao\contao\calendar-bundle\contao\classes\Events.php:97)
  at Contao\Events->getAllEvents()
     (vendor\contao\contao\calendar-bundle\contao\modules\ModuleEventlist.php:173)
  at Contao\ModuleEventlist->compile()
     (vendor\contao\contao\core-bundle\contao\modules\Module.php:214)
  at Contao\Module->generate()
     (vendor\contao\contao\calendar-bundle\contao\modules\ModuleEventlist.php:85)
```

(Noticed via [the community](https://community.contao.org/de/showthread.php?88540-RC2-Type-Error-in-Zusammenhang-mit-getAllEvents-Hook).)

When implementing https://github.com/contao/contao/pull/8390 I unfortunately missed something. I moved the `getAllEvents()` logic to its own service, so that it can be used by the event list as well as the event feed. However, this logic also contains the [`getAllEvents`](https://docs.contao.org/dev/reference/hooks/getAllEvents/) hook of course - and the hook still passes `$this` as its last argument.

This leads to the aforementioned error as the last argument of the hook is supposed to be a `Contao\Module` instance (e.g. the `Contao\ModuleEventlist`.

This PR fixes that by allowing to pass such an instance to the generator method.

This PR also fixes another issue that was present in all previous Contao versions: events that were added via the `getAllEvents` hook were not included in the RSS feed - because the old feed generation simply did not execute this hook. This PR fixes that by creating a faux instance to be passed, with the settings that would apply to an event list. Yes, this is hacky - but until we rewrite the event modules to modern fragments etc. I would leave this as is.

This PR also fixes another small issue: in the generator we allow the URL generation for an event to fail (could happen if the detail page is not selected in the archive or was otherwise deleted). But the feed did not allow it.

Commits
-------

12304f6b Fix getAllEvents Hook
6c28081f also add the type
90969ec0 Update calendar-bundle/contao/classes/Events.php