Restore the previous `autoFocus` when navigating with Turbo Drive (see #8715)
Description
-----------

### Description

- Fixes #8714

_Basically:_
The controller has been added in Contao `5.3`, pre-turbo and `this.autoFocus` was necessary to handle multiple autoFocusTargets as only the first one should be focused on.

Introcuding Turbo in `5.4` also introduced this bug (that did go unnoticed for over a year) as `this.autoFocus` has never been reset to false anymore.

The reason for not being reset whilst navigating is due to the `scroll-offset-controller` being applied to the root of the page (html). Turbo drive replaces the body, thus never affecting the controller anymore / not being able to disconnect the controller.
Setting it to false within a targetDisconnect should fix this behavior.

Commits
-------

0ae3dd25 Reset the previous `autoFocus` when the target disconnected
4728a548 Do not use a stimulus value