Allow to create a user without a request (see #8726)
Description
-----------

The request check seems to be a leftover from https://github.com/contao/contao/pull/4343 that does not make much sense to me. 

I tried to use Symfony `isGrantedForUser` (https://github.com/symfony/symfony/pull/48142) but that requires a `UserInterface`. If I have the `MemberModel`, I would nee this method regardless of having a request or not.

Commits
-------

763aec29 Allow to create user without a request