Only duplicate fragment services if needed (see #8717)
Description
-----------

Improves https://github.com/contao/contao/issues/8255.

We now only duplicate the service if it has multiple `AsContentElement` or `AsFrontendModule` attributes. This means that in case you only add one of them, there will also only be one service (and if you have a `AsHook` attribute, this one will be called only once).

However, as soon as you try to do this with e.g. two `AsContentElement` tags, you will have 3 calls for the hook (1 on the original service and 2 on the duplicated children).

Commits
-------

33b486a1 Only duplicate fragment services if needed
9011f98f Apply changes from review