Do not check the page type when adding child pages (see #8814)
Description
-----------

Currently, if a user does not have access to a page type (e.g. "root"), they cannot create a new child page of any type. We do have to check edit & hierarchy permissions on the _parent_ page (because that is inherited), but a user does not have to have access to a type to create a child.

Commits
-------

015fe04b Do not check page type to add child pages