Fix the template module class (see #8751)
Description
-----------

Added in https://github.com/contao/contao/pull/5310

The module template extends a content element template, which leads to a template class of `content-template` instead of `module-template`.

I'm not sure we need to fix this, it would change the markup for existing users. But I thought we should at least discuss this.
An alternative option would be to just add the `module-template` class so it would have both `content-template module-template`.

Commits
-------

98246cf3 Fix class of template module