Hide the content element preview if it's empty (see #8866)
Description
-----------

Previously, the `border-top` separating a content element's label and operations from the preview was added to `.cte_preview`. But since https://github.com/contao/contao/pull/8518 there is an additional `.cte_content` wrapper and the border was transferred to that. However, this causes the border to also be visible, if the content element has no preview at all (e.g. if the class is missing or if the fragment simply renders no content in the back end), because only the `.cte_preview` wrapper will be set to `display: none`, if it's empty. This causes two borders to be visible at the bottom of the element - one from the outer wrapper and one from `.cte_content`.

This PR fixes that by not outputting the `.cte_content` element at all, if its preview is deemed empty.

Commits
-------

80b17a67 Fix border on empty content element preview
c61e986a remove comments
d0fa5cf7 reverse condition
b6c1d826 remove unused CSS class