Fix the main headline styling in regular back end controllers (see #8858)
Description
-----------

If you have a back end controller and set the `main_headline`, the styling will be slightly off, compared to other DataContainer based back end modules. For example:

<img width="383" src="https://github.com/user-attachments/assets/7c7ac9e4-c162-4ab6-a8b0-4a36bba0e7e6">

\
You can see this by switching back and forth between the _Template Studio_ and any other back end module. The reason for this is that we set a `line-height: 22px` - but only for the `<span>` within, which will (by default) only be present if a breadcrumb is rendered as the main headline.

This PR fixes that by simply moving the `line-height` up to the parent, i.e. `#main_headline` itself.

Commits
-------

9f2f278f Fix main headline styling