Do not render the SERP preview on POST requests (see #8941)
Description
-----------

Fixes #8587

The SERP preview should never be generated for post request to prevent issues with inconsistent data. The only case where the post-generated SERP widget would show up is if the form has errors and the page is not reloaded.

Commits
-------

b16a9831 Do not render the serp preview on POST requests
1531da4f Use `tl_info` instead of `tl_gerror`

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>