Render both content elements and front end modules in the Twig layout (see #8783)
Description
-----------

See #8705

> Content elements are not rendered in the layout because we use Twig's frontend_module method to render the layout module IDs

The change to allow content elements in the layouts came after the initial design of the Twig layouts. To solve the issue, I moved the `frontend_module/module_group.html.twig` template to `layout/_element_group.html.twig`. In there, we are now either using the `content_element()` or `frontend_module()` function depending on the elements type.

Commits
-------

0332e6c9 render both content elements and frontend modules