Ignore web profiler requests in the search index and fix the back end request regex (see #9015)
Description
-----------

Currently the `SearchIndexMessage` is always dispatched for the `/_wdt/{token}` requests of the Symfony profiler in debug mode. This PR fixes that by skipping such URLs.

This PR also fixes 2 issues in the regex for the Contao back end request. First, it had a double slash in it - as `$this->contaoBackendRoutePrefix` already contains a leading slash.

Secondly, if you happen to have a regular page in the frontend that has an alias starting with `contao`, e.g. `/contao-is-awesome`, then that page would never be indexed.

Commits
-------

c132fe96 Skip web profiler requests for search index
38d85449 Remove leading slashes