Add templates for `DC_Table::editAll()` and  `DC_Table::overrideAll()` (see #9005)
Description
-----------

Followup to #8981

This PR…
* moves the HTML out of the `editAll()` and  `overrideAll()` methods
* refactors the "edit" template added in #8981, so that there can be a common base template
* fixes a small bug where error labels where shown twice (details below).

## Review notes
### Template refactoring
I moved some parameter assignments that are global and now used in multiple methods to the `render()` helper. These are: `is_upload_form`, `form_onsubmit` and `error`. This also made the array parameter `form` obsolete in `edit()`.

Because we now have quite some templates specific to editing, I moved all of them to a sub directory `edit`.

Because most of the code is repeated multiple times for edit/edit all/override all (near identical), there now is a common `backend/data_container/table/edit/_edit.html.twig` base template. Everything is wrapped in a `div` containing the jump targets controller and navigation outlet. Here is a small change in the generated output: If the jump target navigation for "edit multiple" (added in https://github.com/contao/contao/pull/8480) was not shown because there were too many items, the surrounding div was also not rendered. Now, we're always rendering the same markup for all views and just conditionally add the respective Stimulus attributes.

### Splitting the flow
Both edit/override all are using two completely independent forms: one to render a selection of available fields and then a second for the edit mask itself. For the first part we are now rendering a common `backend/data_container/table/edit/select_fields.html.twig` template for both operations. 

### Duplicate error messages
In the old code we would output an error message inside the form (see L2515, L2846) like so:
```php
$return .= $this->noReload ? '<p class="tl_error">' . $GLOBALS['TL_LANG']['ERR']['submit'] . '</p>' : '';
```
However, the same error message is then added above the form (shortly before the string is returned), leading to something like this: 
<img width="1045" height="589" alt="image" src="https://github.com/user-attachments/assets/5e674e7d-1a50-4752-baf4-2650fefebfd7" />

I therefore removed the `p` tag inside the form.  By the way: this inconsistency was very easy to spot while abstracting the templates. So first win here. :smile: 

## Future refactoring
After splitting the flow (see above), the code that compiles the selectable fields is now identical in `editAll()` and `overrideAll()`. Also the generation of widget group trees in `edit()` and `editAll()` is _very_ similar. So this should probably be refactored into methods.

Also, the two flows look like this:
```php
// Step 2: edit form
if($fieldsSelected) {
  // […]
  return $this->render('edit/multiple', $parameters);
}

// Step 1: select fields form
// […]
return $this->render('edit/select_fields', $parameters);
```
But now, we could reverse the if condition and have step 1 above step 2, which would move the actual edit part further left (+ with the refactoring from above, the condition body would become a single line).

/cc @diekatrin

Commits
-------

1b7dab1b add templates for edit and override all and refactor existing edit te…
8e174ea8 Merge remote-tracking branch 'origin/5.x' into dc_templates/edit_all
87335ec6 remove comment