Fix accessing a method that potentially does not exist (see #9080)
Description
-----------

Fixes https://github.com/contao/contao/issues/9079

The dependency checker correctly detected, that we're using a class that we don't require. But we ignored it because we do a `class_exists()` check. This is fine but we forgot to also ensure the `method_exists()` on the object.

Commits
-------

a704fed7 Fix accessing method that potentially does not exist