Show the list of changed files in the Twig CS Fixer (see #9143)
Description
-----------

I got annoyed, that the Twig CS Fixer did not output any changed files when running with `fix`, so I made the relevant changes over at https://github.com/VincentLanglet/Twig-CS-Fixer/pull/422. But turns out, this is already possible when increasing the verbosity level. :smile:  So this is what this PR is doing now.

Commits
-------

07d14d70 add the -v flag, so that changed files will be output