Improve the articles node operation (see #9034)
Description
-----------

The articles operation currently generates a URL that includes the current page ID (because it has a simple `href` on the operation). This PR changes that by manually generating the URL.

I noticed that the URL has a request token, because it does affect the session state. This token was not validated though – so that has been added as well. Instead of actually adding the token to the operation, I switched it to POST mode instead though.

Now:
 - the operation no longer has a URL like `do=articles&pn=7&id=7`
 - the operation is a POST form because it affects state
 - the request token is validated if the URL is used directly.

Commits
-------

708826f1 Improve the articles node operation
eaf474fc CI
8e2f5f03 Merge remote-tracking branch 'upstream/5.x' into feature/articles-ope…